* main:
  Escape a character causing a doxygen error.
  Updated release notes.
  Added a comment block for future reference.
  more syntax fixes
  Update CMakeLists.txt
  CMake: Find HDF5 header we can safely include for other checks
  moving functions and macros to new file, lowercase things
  Update release notes.
  lowercase
  lower case
  lowercase
  moving functions and macros to a file
  CMake: Add support for UNITY_BUILD
  Change signature of `nczm_sortenvv` to take `size_t`
  Try to more consistently use `size_t` for argument of `nclistlength`
  Try to more consistently use `size_t` for `nclistget` index argument
  Include header for `size_t` in `nclist.h`

+++ libdispatch/dhttp.c
@@ -8,6 +8,7 @@
 */
 
 #include "config.h"
+#include <stddef.h>
 #include <stdlib.h>
 #include <stdio.h>
 #include <string.h>
@@ -715,7 +716,7 @@ headersoff(NC_HTTP_STATE* state)
 static int
 lookupheader(NC_HTTP_STATE* state, const char* key, const char** valuep)
 {
-    int i;
+    size_t i;
     const char* value = NULL;
     /* Get the content length header */
     for(i=0;i<nclistlength(state->curl.response.headers);i+=2) {
